requirements.txt
cycler==0.10.0
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.6.2
numpy==1.21.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
tifffile==2021.7.30
